% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsSupport.R
\name{checkNameSpace}
\alias{checkNameSpace}
\title{check namespace
Help you to check if you have certain packages and return missing package names}
\usage{
checkNameSpace(packages, quietly = FALSE, from = "")
}
\arguments{
\item{packages}{vector of strings}

\item{quietly}{bool, give you error on fail?}

\item{from}{string, where this package is from like, "CRAN", "GitHub", only
for output message display purpose}
}
\value{
vector strings, of missing package names
}
\description{
check namespace
Help you to check if you have certain packages and return missing package names
}
\examples{
checkNameSpace("ggplot2")
}
