% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{checkSps}
\alias{checkSps}
\title{Pre start SPS checks}
\usage{
checkSps(appDir = getwd())
}
\arguments{
\item{appDir}{where is the app directory root location, default current
working folder}
}
\description{
Pre start SPS checks
}
\examples{
checkSps()
}
