% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_deseq2.R
\name{deg_deseq2}
\alias{deg_deseq2}
\title{Plots a Bar Plot using \code{run_DESeq2} to create a DESeq2 data frame of
differentially expressed genes.}
\usage{
deg_deseq2(countDF, targets, Fold, FDR, cmpset, cmp, plot = T)
}
\arguments{
\item{countDF}{Matrix containing Count data.}

\item{targets}{targets \code{data.frame}}

\item{Fold}{Log Fold Change cut off for filtering.}

\item{FDR}{False Discovery Rate cut off for filtering.}

\item{cmpset}{Number pertaining to index of set of \code{cmp} matrix
desired.}

\item{cmp}{\code{character matrix} where comparisons are defined in two
columns. This matrix should be generated with \code{readComp()} from the
targets file. Values used for comparisons need to match those in the Factor
column of the targets file.}

\item{plot}{If plot = \code{TRUE}, then function will plot a bar plot. If
plot = \code{false}, function will return the DESeq2 \code{data frame}.}
}
\description{
Plots a Bar Plot using \code{run_DESeq2} to create a DESeq2 data frame of
differentially expressed genes.
}
\examples{
## Create DEG dataframe
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- readComp(file=targetspath, format="matrix", delim="-")
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)

## Create plot
deg_deseq2(countDF, targets, Fold = 2, FDR = 10, cmpset = 1, cmp, plot = T)
}
