% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R
\name{devComponents}
\alias{devComponents}
\title{Create template tabs and servers}
\usage{
devComponents(element, shared = NULL)
}
\arguments{
\item{element}{choose from "ui_menu_df", "ui_menu_plot","ui_tab_df",
"ui_tab_plot", "server"}

\item{shared}{use only when \code{element} is 'server'}
}
\value{
ui_xx returns html tags, server will return a server function
}
\description{
generate UI and server part for template components if \code{dev}
option is TRUE. See ui.R and server.R for example. Normally there is no need
to change code of this function ui or server scripts.
}
