% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{dynamicFile}
\alias{dynamicFile}
\title{Dynamically generate Shiny file selection component}
\usage{
dynamicFile(
  id,
  title = "Select your file:",
  label = "Browse",
  icon = NULL,
  style = "",
  multiple = FALSE
)
}
\arguments{
\item{id}{element ID}

\item{title}{element title}

\item{label}{upload button label}

\item{icon}{button icon, only works for \code{local} mode}

\item{style}{additional button style, only works for \code{local} mode}

\item{multiple}{multiple files allowed}
}
\value{
a Shiny upload component
}
\description{
Depending on the "mode" in SPS options, this function renders
a similar UI components but behaves differently on server.
\code{local} mode will not copy file, directly
use a path pointer, \code{server} mode upload file and store in temp. Expect
similar behavior as \link[shiny:fileInput]{shiny::fileInput}.
}
\examples{
if(interactive()){
    library(shinyjs)
    # change to 'local' to see the difference
    options(sps = list(mode='server'))
    ui <- fluidPage(
        useShinyjs(),
        dynamicFile("getFile"),
        textOutput("txt_file")
    )

    server <- function(input,output,session){
        runjs('$(".sps-file input").attr("readonly", true)')
        myfile <- dynamicFileServer(input,session, id = "getFile")
        observe({
            print(myfile())
        })
    }
    shinyApp(ui = ui, server = server)
}
}
