% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{emptyIsFalse}
\alias{emptyIsFalse}
\title{empty things and FALSE will return FALSE}
\usage{
emptyIsFalse(x)
}
\arguments{
\item{x}{expression}
}
\value{
\code{NA}, \code{""}, \code{NULL}, \code{length(0)}, \code{nchar == 0} and \code{FALSE} will return
\code{FALSE}, otherwise \code{TRUE}.
}
\description{
empty things and FALSE will return FALSE
}
\examples{
emptyIsFalse(NULL)
emptyIsFalse(NA)
emptyIsFalse("")
}
