% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlots.R
\name{exploreDDSplot}
\alias{exploreDDSplot}
\title{exploreDDSplot}
\usage{
exploreDDSplot(
  countMatrix,
  targets,
  cmp = cmp[[1]],
  preFilter = NULL,
  sample,
  plotly = FALSE,
  savePlot = FALSE,
  filePlot = NULL
)
}
\arguments{
\item{countMatrix}{\code{date.frame} or \code{matrix} containing raw read counts}

\item{targets}{targets \code{data.frame}}

\item{cmp}{\verb{character matrix} where comparisons are defined in two columns.
This matrix should be generated with the \code{\link[systemPipeR:readComp]{systemPipeR::readComp()}} function from the targets file.
Values used for comparisons need to match those in the \code{Factor} column of the targets file.}

\item{preFilter}{allows removing rows in which there are very few reads.
Accepts a numeric value with the minimum of total reads to keep. Default is \code{NULL}.}

\item{sample}{a character vector of two samples. Could be specified the \code{Factor}
column name of the targets file or the position of the column in the \code{countMatrix}.}

\item{plotly}{logical: when \code{FALSE} (the default), the ggplot2 plot will be returned.}

\item{savePlot}{logical: when \code{FALSE} (the default), the plot will not be saved.
If \code{TRUE} the plot will be saved, and requires the \code{filePlot} argument.}

\item{filePlot}{file name where the plot will be saved. For more information, please consult the
\code{ggsave} function.}
}
\value{
returns an object of \verb{ggplot2 plot}.
}
\description{
Scatterplot of transformed counts from two samples or grid of all samples
}
\examples{
library(systemPipeR)
## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)
## Plot
exploreDDSplot(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, sample=c(3,4))
exploreDDSplot(countMatrix, targets, cmp=cmp[[1]], sample=c("M1A", "M1B"), save = TRUE,
             filePlot = "transf_deseq2.pdf")
}
