% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlots.R
\name{exploreDSS}
\alias{exploreDSS}
\title{exploreDSS
Convenience wrapper function to transform raw read counts using the \code{DESeq2}
package transformations methods. The input file has to contain all the
genes, not just differentially expressed ones.}
\usage{
exploreDSS(
  countMatrix,
  targets,
  cmp = cmp[[1]],
  preFilter = NULL,
  transformationMethod = "raw"
)
}
\arguments{
\item{countMatrix}{\code{date.frame} or \code{matrix} containing raw read counts.}

\item{targets}{targets \code{data.frame}.}

\item{cmp}{\code{character matrix} where comparisons are defined in two columns.
This matrix should be generated with the \code{readComp()} function from the targets file.
Values used for comparisons need to match those in the \code{Factor} column of the targets file.}

\item{preFilter}{allows removing rows in which there are very few reads.
Accepts a numeric value with the minimum of total reads to keep. Default is \code{NULL}.}

\item{transformationMethod}{a character string indicating which transformation
method it will be used on the raw read counts. Supported methods include
\code{rlog} and \code{vst} using the \code{DESeq2} package or default \code{raw}
for no data transformation.}
}
\value{
returns an object of class \code{DESeqTransform}.
}
\description{
exploreDSS
Convenience wrapper function to transform raw read counts using the \code{DESeq2}
package transformations methods. The input file has to contain all the
genes, not just differentially expressed ones.
}
\details{
Note that you can use the resulting transformed values in the
\code{transformationMethod} argument only for visualization and clustering,
not for differential expression analysis which needs raw counts. Users are
strongly encouraged to consult the \code{DESeq2} vignette for more detailed
information on this topic and how to properly run \code{DESeq2} on data sets
with more complex experimental designs.
}
\examples{
library(systemPipeR)
## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)
## Run
exploredss <- exploreDSS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="raw")

}
\references{
For more details on \code{DESeq2}, please consult the following
page: http://bioconductor.org/packages/release/bioc/html/DESeq2.html
For more details on \code{targets} file definition, please consult the following
page: http://www.bioconductor.org/packages/release/bioc/vignettes/systemPipeR/inst/doc/systemPipeR.html#25_structure_of_targets_file
}
\author{
Daniela Cassol
}
