% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsSupport.R
\name{findTabInfo}
\alias{findTabInfo}
\title{Find tab information from tabs.csv}
\usage{
findTabInfo(tabnames = NULL, type = NULL)
}
\arguments{
\item{tabnames}{vector of strings, tab names you want to get}

\item{type}{tab type, one of : core, wf, data, vs}
}
\value{
a list contains `tab_labels`, `hrefs` reference, `images`
}
\description{
Find tab information from tabs.csv
}
\examples{
tabnames <- c("wf_wf", "d", "sas")
findTabInfo(tabnames)
}
