% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsSupport.R
\name{findTreeParent}
\alias{findTreeParent}
\title{find parent steps of from output of jsTree}
\usage{
findTreeParent(step_names)
}
\arguments{
\item{step_names}{}
}
\value{
vector strings of major and minor step numbers
}
\description{
find parent steps of from output of jsTree
}
\examples{
step_name <- c("1.1.1", "2.2.2")
findTreeParent(step_name)
}
