% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R
\name{genGallery}
\alias{genGallery}
\title{generate gallery by only providing tab names}
\usage{
genGallery(
  tabnames = NULL,
  Id = NULL,
  title = "Gallery",
  type = NULL,
  title_color = "#0275d8",
  image_frame_size = 3
)
}
\arguments{
\item{tabnames}{tab names, string vector
will be included}

\item{Id}{div ID}

\item{title}{gallery title}

\item{type}{filter by tab type, then tabnames will be ignored: core, wf, data, vs}

\item{title_color}{title color}

\item{image_frame_size}{integer, 1-12}
}
\value{
gallery div
}
\description{
generate gallery by only providing tab names
}
\examples{
library(shiny)
ui <- fluidPage(
    genGallery(c("tab_a", "tab_b"))
)
server <- function(input, output, session) {

}
shinyApp(ui, server)
}
