% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{genHrefTable}
\alias{genHrefTable}
\title{generate a table of lists of hyper reference buttons by using tab config file}
\usage{
genHrefTable(
  rows,
  Id = NULL,
  title = "A Table of list of tabs",
  text_color = "#0275d8",
  ...
)
}
\arguments{
\item{rows}{a list of rows, each item name in the list will be the row name,
each item is a vector of tab names}

\item{Id}{element ID}

\item{title}{table title}

\item{text_color}{text color for table}

\item{...}{any additional args to the html element, like class, style ...}
}
\description{
generate a table of lists of hyper reference buttons by using tab config file
}
