% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{hexLogo}
\alias{hexLogo}
\alias{hexPanel}
\title{Hexagon logo and logo panel}
\usage{
hexLogo(
  id,
  title = "",
  hex_img,
  hex_link = "",
  footer = "",
  footer_link = "",
  x = "-10",
  y = "-20"
)

hexPanel(
  id,
  title,
  hex_imgs,
  hex_links = NULL,
  hex_titles = NULL,
  footers = NULL,
  footer_links = NULL,
  xs = NULL,
  ys = NULL
)
}
\arguments{
\item{id}{input ID}

\item{title}{title of the logo, display on top of logo or title of logo panel
displayed on the left}

\item{hex_img}{single value of \code{hex_imgs}}

\item{hex_link}{single value of \code{hex_links}}

\item{footer}{single value of \code{footers}}

\item{footer_link}{single value of \code{footer_links}}

\item{x}{character of number, X offset, e.g. "-10" instead of -10L}

\item{y}{character of number, Y offset}

\item{hex_imgs}{a character vector of logo image source, can be online or
local, see details}

\item{hex_links}{a character vector of links attached to each logo, if not
\code{NULL}, must be the same length as \code{hex_imgs}}

\item{hex_titles}{similar to \code{hex_links}, titles of each logo}

\item{footers}{a character vector of footer attached to each logo}

\item{footer_links}{a character vector of footer links, if not \code{NULL},
must be the same length as \code{footers}}

\item{xs}{a character vector X coordinate offset value for each logo image,
default -10}

\item{ys}{Y coordinates offset, must be the same length as \code{xs}, default -20}
}
\value{
HTML elements
}
\description{
Shiny UI widgets to generate hexagon logo(s).
\code{hexLogo()} generates a single hexagon, and \code{hexPanel()}
generates a panel of hex logos
}
\details{
The image in each hexagon is resized to the same size as the hex border
and then enlarged 125\%. You may want to use x, y offset value to change
the image position.

If your image source is local, you need to add your local directory to the
shiny server, e.g. \code{addResourcePath("sps", "www")}. This example add \code{www}
folder under my current working directory as \code{sps} to the server. Then you
can access my images by \code{hex_imgs = "sps/my_img.png"}.

some args in \code{hexPanel} are character vectors, use \code{NULL} for the default
value. If you want to change value but not all of your logos, use \code{""} to
occupy space in the vector. e.g. I have 3 logos, but I only want to add
2 footer and only 1 footer has a link:
\code{footers = c("footer1", "footer2", "")},
\code{footer_links = c("", "https://mylink", "")}. By doing so  \code{footers} and
\code{footer_links} has the same required length.

\code{\link[=useSps]{useSps()}} must be called on Shiny UI to load the required css style.
}
\examples{
if(interactive()){
    ui <- fluidPage(
        useSps(),
        hexPanel(
            "demo1", "DEMO 1:" ,
            rep("https://live.staticflickr.com/7875/46106952034_954b8775fa_b.jpg", 2)
        ),
        hexPanel(
            "demo2", "DEMO 2:" ,
            rep("https://live.staticflickr.com/7875/46106952034_954b8775fa_b.jpg", 2),
            rep("https://www.google.com", 2),
                 c("hex1", "hex2")
        ),
        hexPanel(
            "demo3", "DEMO 3:" ,
            rep("https://live.staticflickr.com/7875/46106952034_954b8775fa_b.jpg", 2),
            footers = c("hex1", "hex2"),
            footer_links = rep("https://www.google.com", 2)
        )
    )
    server <- function(input, output, session) {
    }
    shinyApp(ui, server)
}
}
