% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{hrefTab}
\alias{hrefTab}
\alias{genHrefTab}
\title{Show a list of tabs in buttons}
\usage{
hrefTab(
  Id = NULL,
  title = "A list of tabs",
  title_color = "#0275d8",
  label_text,
  hrefs,
  ...
)

genHrefTab(
  tabnames,
  Id = NULL,
  title = "A bar to list tabs",
  text_color = "#0275d8",
  ...
)
}
\arguments{
\item{Id}{optional}

\item{title}{Item title}

\item{title_color}{title color}

\item{label_text}{individual tab labels}

\item{hrefs}{individual tab links}

\item{...}{other args pass to the html container}

\item{tabnames}{tab names, must have the \code{tab_info} dataframe or
`tab_info.csv` config file in `./config` directory.}
}
\value{
a div element
}
\description{
Show a list of tabs in buttons
}
\details{
`hrefTab` can be use for any purpose of shiny.
`genHrefTab` is upper level wrapper of `hrefTab` and should
only be used under systemPipeShiny framework for fast retrieving tab info and
generate the `hrefTab`.
`label_text`, `hrefs` must be the same length
}
\examples{
ui <- fluidPage(
useSps(),
hrefTab(label_text = c("Bar Plot", "PCA Plot", "Scatter Plot"),
        hrefs = c("https://google.com/", "", ""))
)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
