% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlots.R
\name{MAplot}
\alias{MAplot}
\title{MAplot}
\usage{
MAplot(
  exploredds,
  lfcShrink = FALSE,
  padj.cutoff = 0.05,
  plotly = FALSE,
  savePlot = FALSE,
  filePlot = NULL
)
}
\arguments{
\item{exploredds}{object of class \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}, generated from \code{exploreDDS} function.}

\item{lfcShrink}{logiacal. If \code{TRUE} adds shrunken log2 fold changes (LFC) to the object.}

\item{padj.cutoff}{filter cutoffs for the p-value adjusted.}

\item{plotly}{logical: when \code{FALSE} (default), the \code{ggplot2} plot will be returned.
\code{TRUE} option returns the \code{plotly} version of the plot.}

\item{savePlot}{logical: when \code{FALSE} (default), the plot will not be saved.
If \code{TRUE} the plot will be saved, and requires the \code{filePlot} argument.}

\item{filePlot}{file name where the plot will be saved. For more information, please consult the
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function.}
}
\value{
returns an object of \code{ggplot} or \code{plotly} class.
}
\description{
This function plots log2 fold changes (y-axis) versus the mean
of normalized counts (on the x-axis). Statistically significant features are colored.
}
\examples{
## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)
## Plot
exploredds <- exploreDDS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="raw")
MAplot(exploredds, plotly = FALSE)
MAplot(exploredds, plotly = TRUE)
}
