% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsSupport.R
\name{msg}
\alias{msg}
\title{SPS terminal message}
\usage{
msg(msg, level = "INFO", .other_color = "white")
}
\arguments{
\item{msg}{a character string of message or a vector of character strings,
each item in the vector presents one line of words}

\item{level}{typically, one of "INFO", "WARNING", "ERROR", lower case OK.
Other levels ok.}

\item{.other_color}{hex color code or named colors, when levels are not in
"INFO", "WARNING", "ERROR", this value will be used}
}
\value{

}
\description{
If `crayon` is installed, the message will be colorful.
"INFO" level spawns `message`, "WARNING" is `warning`, "ERROR" spawns `stop`,
other levels use `cat`
}
\examples{
msg("this is info")
msg("this is warning", "warning")
msg("this is error", "error")
msg("this is other", "error")
}
