% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsNewtab.R
\name{newTabData}
\alias{newTabData}
\title{Title}
\usage{
newTabData(
  tab_id = "data_id1",
  tab_displayname = "Data Tab Title",
  desc = "default",
  pkgs = list(cran_pkg = c("base"), bioc_pkg = c(""), github = c("")),
  common_validation = spsValidate({     "pass" }, "common"),
  prepro_methods = list(makePrepro("nothing", "do nothing"), makePrepro("md1",
    "method1", vd_expr = {     nrow(data_filtered) > 1 })),
  app_path = getwd(),
  out_folder_path = file.path(app_path, "R"),
  eg_path = file.path(app_path, "data", "iris.csv"),
  author = "",
  preview = FALSE,
  reformat = TRUE,
  open_file = TRUE,
  use_string = FALSE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)
}
\arguments{
\item{use_string}{}
}
\value{

}
\description{
Title
}
