% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsNewtab.R
\name{newTabPlot}
\alias{newTabPlot}
\title{Title}
\usage{
newTabPlot(
  tab_id = "plot_id1",
  tab_displayname = "Plot Tab Title",
  desc = "default",
  img = "",
  plot_expr = plotly::ggplotly(ggplot(mydata$data, aes_string(names(mydata$data)[1],
    names(mydata$data)[2])) + geom_point(aes(color = seq_len(nrow(mydata$data))))),
  pkgs = list(cran_pkg = c("base"), bioc_pkg = c(""), github = c("")),
  plot_data = list(makePlotData("data", "Raw data")),
  plot_out_func = plotly::plotlyOutput,
  plot_render_func = plotly::renderPlotly,
  app_path = getwd(),
  out_folder_path = file.path(app_path, "R"),
  plot_control_ui = tagList(h3("Some plotting options")),
  author = "",
  preview = FALSE,
  use_string = FALSE,
  reformat = TRUE,
  open_file = TRUE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)
}
\arguments{
\item{colorful}{}
}
\value{

}
\description{
Title
}
