% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsServerFunctions.R, R/spsUIcollections.R
\name{pgPaneUpdate}
\alias{pgPaneUpdate}
\alias{pgPaneUI}
\title{A draggable progress panel
Use `pgPaneUI` on UI side and use `pgPaneUpdate` to update it. The UI only
renders correctly inside `shinydashboard` or `shinydashboardPlus`.}
\usage{
pgPaneUpdate(pane_id, pg_id, value, session = getDefaultReactiveDomain())

pgPaneUI(pane_id, titles, pg_ids, title_main = NULL)
}
\arguments{
\item{pane_id}{Progress panel main ID, use `ns` wrap it on `pgPaneUI` but not
on `pgPaneUpdate` if using shiny module}

\item{pg_id}{a character string of ID for the progress you want to update.
Do not use \code{ns(pg_id)} to wrap it on server}

\item{value}{0-100 real number to update the progress you use `pg_id` to
choose}

\item{session}{current shiny session}

\item{titles}{labels to display for each progress, must have the same length
as `pg_ids`}

\item{pg_ids}{a character vector of IDs for each progress. Don't forget
to use `ns` wrap each ID.}

\item{title_main}{If not specified and pane_id contains 'plot', title will be
'Plot Prepare'; has 'df' will be 'Data Prepare', if neither will be
"Progress"}
}
\description{
A draggable progress panel
Use `pgPaneUI` on UI side and use `pgPaneUpdate` to update it. The UI only
renders correctly inside `shinydashboard` or `shinydashboardPlus`.
}
