% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{quiet}
\alias{quiet}
\title{suppress cat print output}
\usage{
quiet(x)
}
\arguments{
\item{x}{function or expression or value assignment expression}
}
\value{
invisible value
}
\description{
suppress cat print output
}
\examples{
quiet(print(1))
quiet(cat(1))
}
