% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{remove_ANSI}
\alias{remove_ANSI}
\title{Remove ANSI color code}
\usage{
remove_ANSI(string)
}
\arguments{
\item{string}{string}
}
\description{
borrowed from crayon package, since this package is just suggested
by SPS, use some code from crayon
}
