% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsNewtab.R
\name{removeSpsTab}
\alias{removeSpsTab}
\title{Title}
\usage{
removeSpsTab(
  tab_id = "none",
  force = FALSE,
  app_path = getwd(),
  multiple = FALSE,
  verbose = spsOption("verbose"),
  colorful = spsOption("use_crayon")
)
}
\arguments{
\item{colorful}{}
}
\value{

}
\description{
Title
}
