% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{renderDesc}
\alias{renderDesc}
\title{Render some collapsible description}
\usage{
renderDesc(id, desc)
}
\arguments{
\item{id}{HTML ID}

\item{desc}{one string in markdown format}
}
\description{
write some text in markdown format and it will help you render to
a collapsible markdown section
}
\examples{
if(interactive()){
    library(systemPipeShiny)
    library(shiny)
    desc <-
        "
    # Some desc
    - xxxx
    - bbbb

    `Some other things`
    > other markdown things
    1. aaa
    2. bbb
    3. ccc
    "
    ui <- fluidPage(
        useSps(),
        renderDesc(id = "desc", desc)
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}
}
