% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSPS.R
\name{resolveOptions}
\alias{resolveOptions}
\title{Resolve SPS options
need to use the sps_options.yaml file}
\usage{
resolveOptions(appDir = getwd())
}
\arguments{
\item{appDir}{default current working folder}
}
\description{
Resolve SPS options
need to use the sps_options.yaml file
}
