% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CLUST.R
\name{run_CLUST}
\alias{run_CLUST}
\title{Plots Hierarchical Clustering dendrogram between samples using a count
dataframe and a targets file.}
\usage{
run_CLUST(countDF, targets, colData, method)
}
\arguments{
\item{countDF}{Matrix object of Count data.}

\item{targets}{targets data.frame}

\item{colData}{Dataframe containing metadata about each sample.}

\item{method}{Normalization method for plot. Options include "raw", "rlog",
"vst".}
}
\description{
Plots Hierarchical Clustering dendrogram between samples using a count
dataframe and a targets file.
}
\examples{
## Create targets file object
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
colData <- data.frame(row.names = targets$SampleName,
            condition = targets$Factor)
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)
## Create Dendrogram plot
run_CLUST(countDF = countDF, targets = targets, colData = colData, method = "raw")
}
