% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GLM.R
\name{run_GLM}
\alias{run_GLM}
\title{Plots Generalized Principal Component Analysis performed on raw counts using a count dataframe and
a targets file.}
\usage{
run_GLM(countDF, targets, colData)
}
\arguments{
\item{countDF}{Matrix object of Count data.}

\item{targets}{Object containing targets.}

\item{colData}{Dataframe containing metadata about each sample.}
}
\description{
Plots Generalized Principal Component Analysis performed on raw counts using a count dataframe and
a targets file.
}
\examples{
## Create targets file object
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
colData <- data.frame(row.names = targets$SampleName,
            condition = targets$Factor)
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)
## Create GLM-PCA plot
run_GLM(countDF = countDF, targets = targets, colData = colData)
}
