% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MDS.R
\name{run_MDS}
\alias{run_MDS}
\title{Plots Multidimensional Scaling based on pairwise distances between samples
using a count dataframe and a targets file.}
\usage{
run_MDS(countDF, targets, colData, method)
}
\arguments{
\item{countDF}{Matrix object of Count data.}

\item{targets}{Object containing targets.}

\item{colData}{Dataframe containing metadata about each sample.}

\item{method}{Normalization method for plot. Options include "rlog", "vst".}
}
\description{
Plots Multidimensional Scaling based on pairwise distances between samples
using a count dataframe and a targets file.
}
\examples{
## Create targets file object
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
colData <- data.frame(row.names = targets$SampleName,
            condition = targets$Factor)
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)
## Create MDS plot
run_MDS(countDF = countDF, targets = targets, colData = colData, method = "rlog")
}
