% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TSNE.R
\name{run_TSNE}
\alias{run_TSNE}
\title{Plots t-Distributed Stochastic Neighbor Embedding between samples using a
count dataframe and a targets file.}
\usage{
run_TSNE(countDF, targets)
}
\arguments{
\item{countDF}{Matrix object of Count data.}

\item{targets}{Object containing targets.}
}
\description{
Plots t-Distributed Stochastic Neighbor Embedding between samples using a
count dataframe and a targets file.
}
\examples{
## Create targets file object
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
colData <- data.frame(row.names = targets$SampleName,
            condition = targets$Factor)
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)
## Create t-SNE plot
run_TSNE(countDF, targets)
}
