% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_volcano.R
\name{run_volcano}
\alias{run_volcano}
\title{Plots  a Volcano Plot from an \code{edgeR} or \code{DESeq2} data frame.}
\usage{
run_volcano(DF, FDR, Fold, comparison)
}
\arguments{
\item{DF}{\code{edgeR} or \code{DESeq2} data frame.}

\item{FDR}{False Discovery Rate cut off for filtering.}

\item{Fold}{Log Fold Change cut off for filtering.}

\item{Comparison}{\code{String} names of samples or conditions that are
compared. Example: "M1-A1"}
}
\description{
Plots  a Volcano Plot from an \code{edgeR} or \code{DESeq2} data frame.
}
\examples{
## Create DEG dataframe
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- readComp(file=targetspath, format="matrix", delim="-")
countfile <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countDF <- read.delim(countfile, row.names=1)
edgeDF <- run_edgeR(countDF=countDF, targets=targets, cmp=cmp[[1]], independent=FALSE, mdsplot="")

## Create Volcano plot
run_volcano(DF = edgeDF,FDR = 10, Fold = 2, comparison = "M1-A1")
}
