% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSPS.R
\name{runSPS}
\alias{runSPS}
\title{Start systemPipeShiny}
\usage{
runSPS(appDir = getwd())
}
\arguments{
\item{appDir}{where is the app directory root location, default current
working folder}
}
\description{
Start systemPipeShiny
}
\examples{
runSPS()
}
