% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsServerFunctions.R
\name{shinyCatch}
\alias{shinyCatch}
\title{Catch  error, warning, message text catcher}
\usage{
shinyCatch(expr, position = "bottom-right", blocking_level = "none")
}
\arguments{
\item{expr}{expression}

\item{position}{toastr position: c("top-right", "top-center", "top-left",}

\item{blocking_level}{if `non_blocking` is FALSE, at what level you want to
block the execution, one of "error", "warning", "message"}

\item{blocking}{If encounter any error, warning or message, continue or
stop}
}
\value{
see description
}
\description{
Catch error, warning, message by a toastr bar on shiny front end
also log the text on backend console. Will return original value if not
blocking at "warning" "message" level, and return `NULL` at error level.
If blocks at `error`, function will be stopped and other code in the same
reactive context will be blocked. If blocks at `warning` level, warning and
error will be blocked; `message` level blocks will 3 levels.
}
