% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsServerFunctions.R
\name{shinyCheckSpace}
\alias{shinyCheckSpace}
\title{check name space in server
check name space and pop up warnings in shiny if package is missing}
\usage{
shinyCheckSpace(
  session,
  cran_pkg = NULL,
  bioc_pkg = NULL,
  github = NULL,
  quietly = FALSE
)
}
\arguments{
\item{session}{shiny session}

\item{cran_pkg}{vector of strings}

\item{bioc_pkg}{vector of strings}

\item{github}{vector of strings, github package must specify user name, c("user1/pkg1", "user2/pkg2")}

\item{quietly}{bool, should progress and error messages be suppressed?}
}
\value{
sweet alert massage
}
\description{
check name space in server
check name space and pop up warnings in shiny if package is missing
}
