% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{sps}
\alias{sps}
\title{SystemPipeShiny app main function}
\usage{
sps(vstabs = "", plugin = "", server_expr = NULL, app_path = getwd())
}
\arguments{
\item{vstabs}{custom visualization tab IDs that you want to display, in a character
vector. Use \code{\link[=spsTabInfo]{spsTabInfo()}}}

\item{plugin}{If you have loaded some SPS plugins, you can specify here as a
character vector, and it will load all plugin tabs to SPS. If you only want
certain tabs from a plugin, specify in \code{vstabs} argument.}

\item{server_expr}{additional top level sever expression you want
to run. This will run after the default server expressions. It means you can
have access to internal server expression objects, like the reactiveValues
object \code{shared}. You can also overwrite other values.}

\item{app_path}{SPS project path}
}
\value{
a list contains the UI and  server
}
\description{
SystemPipeShiny app main function
}
\details{
You must set the project root as working directory for this
function to find required files.
}
\examples{
if(interactive()){
    spsInit()
    sps_app <- sps(
        vstabs = "",
        server_expr = {
            msg("Hello World", "GREETING", "green")
        }
    )
}
}
