% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsLoading.R
\name{spsUI}
\alias{spsUI}
\title{render the app UI
Combine mainUI defined in ui.R and add the loading screen and more}
\usage{
spsUI(mainUI)
}
\arguments{
\item{mainUI}{a normal shiny page ui}
}
\value{
a `fluidPage`
}
\description{
render the app UI
Combine mainUI defined in ui.R and add the loading screen and more
}
\examples{
ui <- fluidPage()
server <- function(input, output, session) {}
mainUI <- shinyApp(ui, server)
spsUI(mainUI)
}
