% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreServer.R
\name{spsWarnings}
\alias{spsWarnings}
\title{Warning toast under some options when app starts}
\usage{
spsWarnings(session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Warning toast under some options when app starts
}
