% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsSupport.R
\name{step2listTree}
\alias{step2listTree}
\title{Title Take steps output from subsetRmd and change to a nested list structure
Data prepare for ShinyTree}
\usage{
step2listTree(t_lvl, t_text, start_lvl = 0)
}
\arguments{
\item{t_lvl}{positive integers, vector, levels of all title levels in Rmd}

\item{t_text}{character strings, vector, text of titles}

\item{start_lvl}{integer, default value is 0, but default level is 1 (0 + 1). level to start to create list}
}
\value{
a nested list
}
\description{
Title Take steps output from subsetRmd and change to a nested list structure
Data prepare for ShinyTree
}
\examples{
library(shiny)
library(shinyTree)
tree = step2listTree(t_lvl, t_text)
str(tree)

tree_names = names(unlist(tree))

ui = shinyUI(
    pageWithSidebar(
        mainPanel(
            shinyTree("tree", stripes = TRUE, multiple = FALSE, animation = FALSE)
        )
    ))
server = shinyServer(function(input, output, session) {
    output$tree <- renderTree({
        tree
    })
})
shinyApp(ui, server)
}
