% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{tabTitle}
\alias{tabTitle}
\title{h2 title with bootstrap info color}
\usage{
tabTitle(title, ...)
}
\arguments{
\item{title}{title text}

\item{...}{other attributes and children to this DOM}
}
\value{
a h2 level heading with bootstrap4 info color, bt4 color not the
default bt3 info color
}
\description{
h2 title with bootstrap info color
}
\examples{
tabTitle("This title")
}
