% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{textInputGroup}
\alias{textInputGroup}
\title{Text input with a button right next to}
\usage{
textInputGroup(textId, btnId, title = "", label = "", icon = "paper-plane")
}
\arguments{
\item{textId}{text box id}

\item{btnId}{action button id}

\item{title}{title of this group}

\item{label}{button text}

\item{icon}{button icon}
}
\value{
a row element
}
\description{
Text input with a button right next to
}
\examples{
if(interactive()){
    library(systemPipeShiny)

    ui <- fluidPage(
        useSps(),
        textInputGroup("id1", "id2")
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}
}
