% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{uiExamples}
\alias{uiExamples}
\title{Example UI elements for plotting}
\usage{
uiExamples(ns)
}
\arguments{
\item{ns}{namespace function}
}
\value{

}
\description{
return some example UI elements can be toggled on plotting
}
