% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{viewSpsDefaults}
\alias{viewSpsDefaults}
\title{Print systemPipeShiny default Options}
\usage{
viewSpsDefaults(app_path = getwd())
}
\arguments{
\item{app_path}{where is the app directory}
}
\value{
cat to console the default options
}
\description{
Make sure you created the app folder and has config.yaml
in config folder
}
\examples{
if(interactive()){
    spsInit()
    viewSpsDefaults()
}
}
