% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{wfPanel}
\alias{wfPanel}
\title{Workflow progress tracker UI}
\usage{
wfPanel()
}
\description{
call it on top level UI not inside a module. Call this function
only once. Do not repeat this function.
}
\examples{
wfPanel()
}
