% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{clearableTextInput}
\alias{clearableTextInput}
\title{A clearable text inputInput control}
\usage{
clearableTextInput(inputId, label, value = "", placeholder = "")
}
\arguments{
\item{inputId}{ID}

\item{label}{text label above}

\item{value}{default value}

\item{placeholder}{place holder text}
}
\value{
a div
}
\description{
An UI component with a "X" button in the end to clear the entire
entered text. It works the same as \code{Textinput}.
}
\details{
must be call \code{\link[=useSps]{useSps()}} at head to load css and js
}
\examples{
if(interactive()){

    ui <- fluidPage(
        useSps(),
        clearableTextInput("input1", "This is a input box")
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}
}
