% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{clearableTextInput}
\alias{clearableTextInput}
\title{Input component with a "X" button in the end to clear the entire typed text
work the same as `Textinput`
must be call `useSps` at header}
\usage{
clearableTextInput(inputId, label, value = "", placeholder = "")
}
\arguments{
\item{inputId}{ID}

\item{label}{text label above}

\item{value}{default value}

\item{placeholder}{place holder text}
}
\value{
a div
}
\description{
Input component with a "X" button in the end to clear the entire typed text
work the same as `Textinput`
must be call `useSps` at header
}
\examples{
library(shiny)

ui <- fluidPage(
    useSps(),
    clearableTextInput("input1", "This is a input box")
)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
