% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{emptyIsFalse}
\alias{emptyIsFalse}
\title{empty things and FALSE will return FALSE}
\usage{
emptyIsFalse(x)
}
\arguments{
\item{x}{expression}
}
\description{
empty things and FALSE will return FALSE
}
\examples{
emptyIsFalse(NULL)
emptyIsFalse(NA)
emptyIsFalse("")
}
