% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{findTabInfo}
\alias{findTabInfo}
\title{Find tab information from tabs.csv
If \code{type} is not empty, \code{tabnames} will be ignored}
\usage{
findTabInfo(tabnames = NULL, type = NULL)
}
\arguments{
\item{tabnames}{vector of strings, tab names you want to get}

\item{type}{tab type and sub type, one of: core, wf, vs, data, plot}
}
\value{
a list contains \code{tab_labels}, \code{hrefs} reference, \code{images}
}
\description{
Find tab information from tabs.csv
If \code{type} is not empty, \code{tabnames} will be ignored
}
\examples{
tabnames <- c("wf_wf", "d", "sas")
findTabInfo(tabnames)
}
