% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R
\name{genGallery}
\alias{genGallery}
\title{generate gallery by only providing tab names}
\usage{
genGallery(
  tab_ids = NULL,
  Id = NULL,
  title = "Gallery",
  type = NULL,
  title_color = "#0275d8",
  image_frame_size = 3
)
}
\arguments{
\item{tab_ids}{tab IDs, string vector
will be included}

\item{Id}{div ID}

\item{title}{gallery title}

\item{type}{filter by tab type, then tab_ids will be ignored: core, wf, data, vs}

\item{title_color}{title color}

\item{image_frame_size}{integer, 1-12}
}
\value{
gallery div
}
\description{
generate gallery by only providing tab names
}
\examples{
if(interactive()){
    library(systemPipeShiny)
    spsInit()
    ui <- fluidPage(
        genGallery(c("plot_template")),
        genGallery(type = "plot")
    )
    server <- function(input, output, session) {

    }
    shinyApp(ui, server)
}
}
