% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R
\name{genHrefTable}
\alias{genHrefTable}
\title{generate a table of lists of hyper reference buttons by using tab config file}
\usage{
genHrefTable(
  rows,
  Id = NULL,
  title = "A Table to list tabs",
  text_color = "#0275d8",
  ...
)
}
\arguments{
\item{rows}{a named list of character vector, each item name in the list
will be the row name, each item is a vector of tab names. Or you can use one
of 'core', 'wf', 'vs', 'data', 'plot' to specify a tab type, so it will find
all tabs matching that type. See \code{tab_info.csv} under \code{config} directory for
type info.}

\item{Id}{element ID}

\item{title}{table title}

\item{text_color}{text color for table}

\item{...}{any additional args to the html element, like class, style ...}
}
\description{
generate a table of lists of hyper reference buttons by using tab config file
}
\details{
For \code{rows}, there are some specially reserved characters
for type and sub types. If indicated, it will return a list of tabs matching
the indicated tabs instead of searching individual tab names. These words
include: core, wf, vs, data, plot.
}
\examples{
if(interactive()){
    library(systemPipeShiny)
    spsInit()
    rows <- list(wf1 = c("core_canvas", "core_about"), wf2 =  "data")
    ui <- fluidPage(
        genHrefTable(rows)
    )
    server <- function(input, output, session) {

    }
    shinyApp(ui, server)
}
}
