% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R, R/spsUIcollections.R
\name{genHrefTab}
\alias{genHrefTab}
\alias{hrefTab}
\title{Show a list of tabs in buttons}
\usage{
genHrefTab(
  tabnames,
  Id = NULL,
  title = "A bar to list tabs",
  text_color = "#0275d8",
  ...
)

hrefTab(
  Id = NULL,
  title = "A list of tabs",
  title_color = "#0275d8",
  label_text,
  hrefs,
  ...
)
}
\arguments{
\item{tabnames}{tab names, must have the \code{tab_info} dataframe}

\item{Id}{optional}

\item{title}{Item title}

\item{text_color}{Table text color}

\item{...}{other args pass to the html container}

\item{title_color}{title color}

\item{label_text}{individual tab labels}

\item{hrefs}{individual tab links}
}
\value{
a div element
}
\description{
Show a list of tabs in buttons
}
\details{
\code{hrefTab} can be use for any purpose of shiny.
\code{genHrefTab} is upper level wrapper of \code{hrefTab} and should
only be used under systemPipeShiny framework for fast retrieving tab info and
generate the \code{hrefTab}. To use \code{genHrefTab}, the \code{tab_info.csv} config file
must be in \code{./config} directory.
\code{label_text}, \code{hrefs} must be the same length
}
\examples{
ui <- fluidPage(
useSps(),
hrefTab(label_text = c("Bar Plot", "PCA Plot", "Scatter Plot"),
        hrefs = c("https://google.com/", "", ""))
)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
