% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R, R/spsUIcollections.R
\name{genHrefTab}
\alias{genHrefTab}
\alias{hrefTab}
\title{Display a list of links in a row of buttons}
\usage{
genHrefTab(
  tab_ids,
  Id = NULL,
  title = "A bar to list tabs",
  text_color = "#0275d8",
  ...
)

hrefTab(
  Id = NULL,
  title = "A list of tabs",
  title_color = "#0275d8",
  label_text,
  hrefs,
  ...
)
}
\arguments{
\item{tab_ids}{tab names, must have the \code{tab_info} dataframe}

\item{Id}{optional element ID}

\item{title}{Item title}

\item{text_color}{Table text color}

\item{...}{other arguments to be passed to the html element}

\item{title_color}{title color}

\item{label_text}{individual tab labels}

\item{hrefs}{individual tab links}
}
\value{
a HTML element
}
\description{
\code{hrefTab} can be use for any purpose of shiny.
\code{genHrefTab} is upper level wrapper of \code{hrefTab} and should
only be used under SPS framework for fast retrieving tab info and
generate the \code{hrefTab}. To use \code{genHrefTab}, the \code{tab_info.csv} config file
must be there in \code{config} directory.
\code{label_text}, \code{hrefs} must be the same length
}
\details{
\emph{genHrefTab} require a SPS project and the config/tabs.csv file.
If you want to use hrefTab outside a SPS project just use \emph{hrefTab}
}
\examples{
if(interactive()){
    ui <- fluidPage(
        useSps(),
        hrefTab(label_text = c("Bar Plot", "PCA Plot", "Scatter Plot"),
                hrefs = c("https://google.com/", "", ""))
    )

    server <- function(input, output, session) {

    }
    shinyApp(ui, server)
}
}
