% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{hrefTable}
\alias{hrefTable}
\title{A table of lists of hyper reference buttons
\code{item_titles}, \code{item_labels}, \code{item_hrefs} must have the same length
nth item in \code{item_labels}, \code{item_hrefs} must have the same length}
\usage{
hrefTable(
  Id = NULL,
  title = "A Table of list of tabs",
  text_color = "#0275d8",
  item_titles,
  item_labels,
  item_hrefs,
  ...
)
}
\arguments{
\item{Id}{optional}

\item{title}{title of this table}

\item{text_color}{text color}

\item{item_titles}{vector string, a vector of titles for table items}

\item{item_labels}{list, a list of button lables in each table item}

\item{item_hrefs}{list, a list of hrefs for each button}

\item{...}{other HTML args}
}
\value{

}
\description{
A table of lists of hyper reference buttons
\code{item_titles}, \code{item_labels}, \code{item_hrefs} must have the same length
nth item in \code{item_labels}, \code{item_hrefs} must have the same length
}
\examples{
library(shiny)

ui <- fluidPage(
    useSps(),
    hrefTable(item_titles = c("workflow 1", "workflow 2"),
              item_labels = list(c("tab 1"), c("tab 3", "tab 4")),
              item_hrefs = list(c("https://www.google.com/"), c("", "")),
    )

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
