% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsNewtab.R
\name{makePlotData}
\alias{makePlotData}
\title{Title}
\usage{
makePlotData(
  dataset_id = "data",
  dataset_label = "Raw data",
  receive_datatab_ids = "data_template",
  vd_expr = spsValidate({     if (is.data.frame(mydata$data))          TRUE     else
    stop("Data xx needs to be a dataframe or tibble") }),
  use_string = FALSE
)
}
\arguments{
\item{use_string}{}
}
\value{

}
\description{
Title
}
