% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsNewtab.R
\name{makePrepro}
\alias{makePrepro}
\title{Title}
\usage{
makePrepro(
  method_id = "md1",
  label = "New method1",
  vd_expr = spsValidate(is.data.frame(data_filtered)),
  pre_expr = {     cat("do nothing")     data_filtered },
  plot_options = "default",
  use_string = FALSE
)
}
\arguments{
\item{use_string}{}
}
\value{

}
\description{
Title
}
