% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{msg}
\alias{msg}
\alias{spsinfo}
\alias{spswarn}
\alias{spserror}
\title{SPS terminal message logging methods}
\usage{
msg(
  msg,
  level = "INFO",
  .other_color = "white",
  info_text = "INFO",
  warning_text = "WARNING",
  error_text = "ERROR"
)

spsinfo(msg, verbose = spsOption("verbose"))

spswarn(msg)

spserror(msg)
}
\arguments{
\item{msg}{a character string of message or a vector of character strings,
each item in the vector presents one line of words}

\item{level}{typically, one of "INFO", "WARNING", "ERROR", lower case OK.
Other levels ok.}

\item{.other_color}{hex color code or named colors, when levels are not in
"INFO", "WARNING", "ERROR", this value will be used}

\item{info_text}{info level text prefix}

\item{warning_text}{warning level text prefix}

\item{error_text}{error level text prefix}

\item{verbose}{bool, default get from sps project options, can be overwritten}
}
\value{
see description
}
\description{
If SPS \code{use_crayon}option is \code{TRUE}, the message will
be colorful.
"INFO" level spawns \code{message}, "WARNING" is \code{warning}, "ERROR" spawns \code{stop},
other levels use \code{cat}.
\code{spsinfo}, \code{spswarn}, \code{spserror} are higher level wrappers of \code{msg}. The
only difference is they have \verb{SPS-} prefix.

\code{spsinfo} has an additional
arg \code{verbose}. This arg works similar to all other \code{verbose} args in this
package, if not specified, it follows the project option, can be be forced to
\code{TRUE} and \code{FALSE}. \code{TRUE} will forcefully generate the msg, and \code{FALSE}
will be no message.
}
\examples{
msg("this is info")
msg("this is warning", "warning")
try(msg("this is error", "error"))
msg("this is other", "error2")
spsinfo("some msg")
spswarn("sps warning")
try(spserror("sps error"))

}
