% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{msg}
\alias{msg}
\title{SPS terminal message}
\usage{
msg(
  msg,
  level = "INFO",
  .other_color = "white",
  info_text = "INFO",
  warning_text = "WARNING",
  error_text = "ERROR"
)
}
\arguments{
\item{msg}{a character string of message or a vector of character strings,
each item in the vector presents one line of words}

\item{level}{typically, one of "INFO", "WARNING", "ERROR", lower case OK.
Other levels ok.}

\item{.other_color}{hex color code or named colors, when levels are not in
"INFO", "WARNING", "ERROR", this value will be used}

\item{info_text}{info level text prefix}

\item{warning_text}{warning level text prefix}

\item{error_text}{error level text prefix}
}
\value{
see description
}
\description{
If SPS \code{use_crayon}option is \code{TRUE}, the message will
be colorful.
"INFO" level spawns \code{message}, "WARNING" is \code{warning}, "ERROR" spawns \code{stop},
other levels use \code{cat}
}
\examples{
msg("this is info")
msg("this is warning", "warning")
try(msg("this is error", "error"))
msg("this is other", "error2")
}
