% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{renderDesc}
\alias{renderDesc}
\title{Render tab description}
\usage{
renderDesc(id, desc)
}
\arguments{
\item{desc}{one string in markdown format}
}
\description{
Render tab description
}
\examples{
desc <-
"
# Some desc
- xxxx
"
renderDesc(desc, id = "desc")
}
