% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{renderDesc}
\alias{renderDesc}
\title{Render some collapsible markdown text}
\usage{
renderDesc(id, desc)
}
\arguments{
\item{id}{HTML ID}

\item{desc}{one string in markdown format}
}
\value{
HTML elements
}
\description{
write some text in markdown format and it will help you
render to a collapsible markdown section on Shiny UI
}
\examples{
if(interactive()){
    library(shiny)
    desc <-
        "
    # Some desc
    - xxxx
    - bbbb

    `Some other things`
    > other markdown things
    1. aaa
    2. bbb
    3. ccc
    "
    ui <- fluidPage(
        useSps(),
        renderDesc(id = "desc", desc)
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}
}
