% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{resolveOptions}
\alias{resolveOptions}
\title{Resolve SPS options
need to use the sps_options.yaml file}
\usage{
resolveOptions(appDir = getwd())
}
\arguments{
\item{appDir}{default current working folder}
}
\description{
Resolve SPS options
need to use the sps_options.yaml file
}
\examples{
options(sps = list(mode = c("asas", "sas"), place = "here", time = c("now", "then"), loading_screen = FALSE))
resolveOptions()
}
