% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsServerCollections.R
\name{shinyCheckPkg}
\alias{shinyCheckPkg}
\title{check name space in server
check name space and pop up warnings in shiny if package is missing}
\usage{
shinyCheckPkg(
  session,
  cran_pkg = NULL,
  bioc_pkg = NULL,
  github = NULL,
  quietly = FALSE
)
}
\arguments{
\item{session}{shiny session}

\item{cran_pkg}{vector of strings}

\item{bioc_pkg}{vector of strings}

\item{github}{vector of strings, github package must specify user name, c("user1/pkg1", "user2/pkg2")}

\item{quietly}{bool, should progress and error messages be suppressed?}
}
\value{
TRUE if pass, sweet alert massage and FALSE if fail
}
\description{
check name space in server
check name space and pop up warnings in shiny if package is missing
}
\examples{
if(interactive()){
    library(systemPipeShiny)
    shinyApp(ui = shinyUI(
        fluidPage(actionButton("haha", "haha"))
    ), server = function(input, output, session) {
        observeEvent(input$haha, shinyCheckPkg(session, cran_pkg = "1",
                                               bioc_pkg = "haha", github = "sdasdd/asdsad"))
    })
}
}
