% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{sps}
\alias{sps}
\title{SystemPipeShiny app start function}
\usage{
sps(vstabs, server_expr = NULL)
}
\arguments{
\item{vstabs}{your custom visualization tabs you want to display}

\item{server_expr}{additional top level sever expression you want to run. This
will run after the default server expressions. It means you can have access to
internal server expression objects, like the reactiveValues object \code{shared}.
You can also overwrite other values.}
}
\description{
SystemPipeShiny app start function
}
\examples{
# on ui.R
sps_app <<- sps(
vstabs = "",
server_expr = {
    msg("Hello World", "GREETING", "green")
}
)
}
