% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{spsInit}
\alias{spsInit}
\title{Create a SystemPipeShiny project}
\usage{
spsInit(
  dir_path = getwd(),
  project_name = glue("SPS_{format(Sys.time(), '\%Y\%m\%d')}"),
  database_name = "sps.db",
  overwrite = FALSE,
  change_wd = TRUE,
  verbose = FALSE,
  open_files = TRUE,
  colorful = TRUE
)
}
\arguments{
\item{dir_path}{path, where do you want to create this project}

\item{project_name}{Your project name}

\item{database_name}{project database name, recommend to use the default name}

\item{overwrite}{bool, overwrite the \code{dir_path}?}

\item{change_wd}{bool, when creation is done, change working directory?}

\item{verbose}{bool, do you want additional message?}

\item{open_files}{bool, If \code{change_wd == TRUE} and you are also in Rstudio,
it will open up global.R, ui.R and server.R for you}

\item{colorful}{bool, should message from this function be colorful?}
}
\value{
creates the project folder
}
\description{
Create a SystemPipeShiny project
}
\examples{
if(interactive()){
    spsInit(change_wd = FALSE)
}
}
