% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreUI.R
\name{spsUI}
\alias{spsUI}
\title{Generate SPS UI}
\usage{
spsUI(tabs_df, tabs_plot)
}
\arguments{
\item{tabs_df}{vs data tab IDs}

\item{tabs_plot}{vs plot tab IDs}
}
\value{
shiny dash page
}
\description{
Generate SPS UI
}
\details{
Workflow tabs and other \code{core} tabs are loaded by default. You can
only optionally choose visualization tabs. See config/tabs.csv for tab info.
}
\examples{
spsUI()
}
