% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUti.R
\name{step2listD3}
\alias{step2listD3}
\title{Create structure for networkD3 object}
\usage{
step2listD3(t_lvl, t_text, start_lvl = 0)
}
\arguments{
\item{t_lvl}{title markdown heading levels, 1-5}

\item{t_text}{title text}

\item{start_lvl}{starting title level}
}
\value{
list
}
\description{
Create structure for networkD3 object
}
\examples{
t_lvl = c(1, 3, 1, 2, 2, 3)
t_text = c('1', '1.1.1', '2', '2.1', '2.2', '2.2.1')
test = step2listD3(t_lvl, t_text)
str(test)
diagonalNetwork(test)
}
