% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{uiExamples}
\alias{uiExamples}
\title{Example UI elements for plotting}
\usage{
uiExamples(ns)
}
\arguments{
\item{ns}{namespace function}
}
\value{
some UI
}
\description{
return some example UI elements can be toggled on plotting.
This functions is only used as a temp solution for template tab to
demonstrate what UI compoents
you could use. Will be removed in later releases as we have a better tab
organizations.
}
\examples{
ui <- fluidPage(useSps(), uiExamples(NS("example")))
server <- function(input, output, session) {}
shinyApp(ui, server)
}
