% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{useSps}
\alias{useSps}
\title{Use SystemPipeShiny javascripts and css style
call it in your head section}
\usage{
useSps()
}
\value{

}
\description{
Use SystemPipeShiny javascripts and css style
call it in your head section
}
\examples{
library(systemPipeShiny)
useSps()
}
