% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsUIcollections.R
\name{useSps}
\alias{useSps}
\title{Use SystemPipeShiny javascripts and css style}
\usage{
useSps()
}
\value{
HTML head
}
\description{
call it in your head section of your shiny UI. Adding this to
the Shiny app UI is required for most SPS shiny widgets. This is required for
using SPS widgets outside of SPS framework. No need to load this function
again if you are working within SPS framework.
}
\examples{
# useSps()
}
