% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSPS.R
\name{ViewSpsDefaults}
\alias{ViewSpsDefaults}
\title{Print systemPipeShiny default Options}
\usage{
ViewSpsDefaults(appDir = getwd())
}
\arguments{
\item{appDir}{where is the app directory}
}
\description{
Print systemPipeShiny default Options
}
\examples{
# Make sure you created the app folder and has config.yaml in config folder
ViewSpsDefaults()
}
