% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCore.R
\name{viewSpsDefaults}
\alias{viewSpsDefaults}
\title{Print SPS default options}
\usage{
viewSpsDefaults(app_path = getwd())
}
\arguments{
\item{app_path}{path, where is the app directory}
}
\value{
cat to console the default options
}
\description{
Make sure you have created the app directory and it
has \emph{config/config.yaml} file
}
\examples{
if(interactive()){
    spsInit(open_files = FALSE)
    viewSpsDefaults()
}
}
