% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsCoreServer.R
\name{wfProgressPanel}
\alias{wfProgressPanel}
\title{Workflow Progress tracker server logic}
\usage{
wfProgressPanel(shared)
}
\arguments{
\item{shared}{the shared object}
}
\value{
reactive renderUI object
}
\description{
use it on the top level server not inside a module. Only
designed for workflow tabs. For other purpose, use \code{pgPaneUI}.
}
\examples{
wfProgressPanel(shared)
}
\seealso{
pgPaneUpdate
}
